Started refactoring of Controller (Home)
