+++ ui/index.html
@@ -11,7 +11,7 @@
     <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
     <link rel="manifest" href="/site.webmanifest" />
   </head>
-  <body class="bg-gray-100">
+  <body class="bg-main-800 text-black/60 @dark:(bg-main-200 text-white/80)">
     <div id="app"></div>
     <script type="module" src="./src/main.ts"></script>
   </body>
