+++ ui/src/Home/NewBoardDialog.svelte
@@ -1,15 +1,15 @@
 <script lang="ts">
-    import BoardEditor from './BoardEditor.svelte';
-    import type { GamezStore } from './store';
+    import BoardEditor from '../BoardEditor.svelte';
+    import type { GamezStore } from '../store';
     import { getContext } from 'svelte';
-    import type { BoardProps, BoardState, PieceDef } from './board';
+    import type { BoardProps, BoardState, PieceDef } from '../board';
     import '@shoelace-style/shoelace/dist/components/dialog/dialog.js';
     import '@shoelace-style/shoelace/dist/components/button/button.js';
     import type SlDialog from '@shoelace-style/shoelace/dist/components/dialog/dialog';
 
     let editPieceDefs = []
     let dialog: SlDialog
-    
+
     const { getStore } :any = getContext('gzStore');
 
     const store:GamezStore = getStore();
@@ -29,7 +29,7 @@
 <sl-dialog style="--width:600px" bind:this={dialog} label="New Game Type"
     on:sl-request-close={(event)=>{
         if (event.detail.source === 'overlay') {
-        event.preventDefault();    
+        event.preventDefault();
   }}}>
     <BoardEditor bind:this={boardEditor}  handleSave={addBoard} cancelEdit={()=>dialog.hide()} />
 </sl-dialog>
