+++ ui/src/app.css
@@ -51,6 +51,7 @@ html, body {
   padding: 0px;
   height: 100%;
   overflow: hidden; /* In general, it's a good idea to have the body take up the full viewport height and have a fixed layout with no scrollbar. Then, you can have a scrollable inner container like .pane-content to handle the overflow of the content. This approach helps keep the page structure clean and allows you to control the scrolling behavior more effectively. */
+  background-image: url('/images/noise.png');
 }
 
 a {
