+++ ui/src/index.js
@@ -1,20 +1,20 @@
 // @ts-nocheck
-export * from './board';
-export * from './gamezStore';
-export * from './boardList';
-export { default as Controller } from './Controller.svelte';
-export { default as BoardEditor } from './BoardEditor.svelte';
-export { default as AvatarIcon } from './AvatarIcon.svelte';
-export { default as AvatarDialog } from './AvatarDialog.svelte';
-export { default as AboutDialog } from './AboutDialog.svelte';
-export { default as ParticipantsDialog } from './ParticipantsDialog.svelte';
-export { default as GamezPane } from './GamezPane';
-export { default as BoardMenu} from './BoardMenu.svelte';
-export { default as Folk} from './Folk.svelte';
-export { default as NewBoardDialog} from './NewBoardDialog.svelte';
-export { default as EditBoardDialog} from './EditBoardDialog.svelte';
-export { default as CardEditor} from './CardEditor.svelte';
-export { default as CardEditor} from './CardEditor.svelte';
-export { default as Toolbar } from './Toolbar.svelte';
-export { default as EmojiIcon } from './icons/EmojiIcon.svelte';
-export { default as LogoIcon } from './icons/LogoIcon.svelte';
+export * from "./board";
+export * from "./gamezStore";
+export * from "./boardList";
+export { default as Controller } from "./Controller.svelte";
+export { default as BoardEditor } from "./BoardEditor.svelte";
+export { default as AvatarIcon } from "./AvatarIcon.svelte";
+export { default as AvatarDialog } from "./AvatarDialog.svelte";
+export { default as AboutDialog } from "./AboutDialog.svelte";
+export { default as ParticipantsDialog } from "./ParticipantsDialog.svelte";
+export { default as GamezPane } from "./GamezPane";
+export { default as BoardMenu } from "./BoardMenu.svelte";
+export { default as Folk } from "./Folk.svelte";
+export { default as NewBoardDialog } from "./Home/NewBoardDialog.svelte";
+export { default as EditBoardDialog } from "./Home/EditBoardDialog.svelte";
+export { default as CardEditor } from "./CardEditor.svelte";
+export { default as CardEditor } from "./CardEditor.svelte";
+export { default as Toolbar } from "./Toolbar.svelte";
+export { default as EmojiIcon } from "./icons/EmojiIcon.svelte";
+export { default as LogoIcon } from "./icons/LogoIcon.svelte";
